requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
Pillow>=10.0.0
tqdm>=4.65.0
streamlit>=1.28.0
plotly>=5.17.0
einops>=0.7.0
tensorboard>=2.14.0
opencv-python>=4.8.0